fix package location
