--- server/release_notes_test/pmm/2.43.0-rel-links.md
@@ -1,3 +1,0 @@
-### PMM 2.42.0
-Welcome to PMM [v2.42](../index.md)
--![!image](../_images/Max_Connection_Limit.png)
